+++ src/low-level/nntp/newsnntp.c
@@ -96,7 +96,7 @@ newsnntp * newsnntp_new(size_t progr_rate, progress_function * progr_fun)
   f->nntp_stream = NULL;
   f->nntp_readonly = FALSE;
 
-  f->nntp_progr_rate = progr_rate;
+  f->nntp_progr_rate = (uint32_t)progr_rate;
   f->nntp_progr_fun = progr_fun;
 
   f->nntp_stream_buffer = mmap_string_new("");
@@ -1814,7 +1814,7 @@ static int parse_response(newsnntp * f, char * response)
 {
   int code;
 
-  code = strtol(response, &response, 10);
+  code = (int)strtol(response, &response, 10);
 
   if (response == NULL) {
     f->nntp_response = NULL;
@@ -1863,15 +1863,15 @@ static int parse_group_info(char * response,
 
   line = response;
 
-  count = strtoul(line, &line, 10);
+  count = (uint32_t)strtoul(line, &line, 10);
   if (!parse_space(&line))
     return FALSE;
 
-  first = strtoul(line, &line, 10);
+  first = (uint32_t)strtoul(line, &line, 10);
   if (!parse_space(&line))
     return FALSE;
 
-  last = strtoul(line, &line, 10);
+  last = (uint32_t)strtoul(line, &line, 10);
   if (!parse_space(&line))
     return FALSE;
 
@@ -1920,11 +1920,11 @@ static clist * read_groups_list(newsnntp * f)
     group_name = line;
     line = p;
 
-    last = strtol(line, &line, 10);
+    last = (uint32_t)strtol(line, &line, 10);
     if (!parse_space(&line))
       continue;
 
-    first = strtol(line, &line, 10);
+    first = (uint32_t)strtol(line, &line, 10);
     if (!parse_space(&line))
       continue;
 
@@ -2135,7 +2135,7 @@ static clist * read_distrib_default_value_list(newsnntp * f)
 
     p = line;
 
-    weight = strtoul(p, &remaining, 10);
+    weight = (uint32_t)strtoul(p, &remaining, 10);
     p = remaining;
     parse_space(&p);
       
@@ -2321,7 +2321,7 @@ static clist * read_xhdr_resp_list(newsnntp * f)
     if (mailstream_is_end_multiline(line))
       break;
     
-    article = strtoul(line, &line, 10);
+    article = (uint32_t)strtoul(line, &line, 10);
     if (!parse_space(&line))
       continue;
